+++ functions.php
@@ -83,13 +83,25 @@ function hrs_login_logo_url_title() {
 	return get_bloginfo( 'name' );
 }
 
-/***** start old ******/
+add_filter( 'spine_get_title', 'hrs_get_page_title' );
+/**
+ * Filters the Spine page title contents.
+ *
+ * Adjusts the formatting and punctuation of the default Spine parent theme
+ * title, which itself replaces the default `wp_title()`,
+ * {@see https://github.com/washingtonstateuniversity/WSUWP-spine-parent-theme/functions.php}
+ *
+ * @since 0.12.0
+ *
+ * @param string $title The built title to filter.
+ * @return string The contents of the `title` element.
+ */
+function hrs_get_page_title( $title ) {
+	$page_title   = wp_title( '-', false, 'right' );
+	$site_title   = get_bloginfo( 'name' );
+	$global_title = ', Washington State University';
 
-add_filter( 'gettext', 'remove_lostpassword_text' );
-/** Remove lost password **/
-function remove_lostpassword_text( $text ) {
-	if ( 'Lost your password?' === $text ) {
-		$text = '';
-	}
-	return $text;
+	$title = $page_title . $site_title . $global_title;
+
+	return $title;
 }